Move pivot files
