+++ web-common/src/features/dashboards/pivot/PivotTable.svelte
@@ -1,0 +1,167 @@
+<script lang="ts">
+  import { writable } from "svelte/store";
+  import {
+    ColumnDef,
+    createSvelteTable,
+    flexRender,
+    getCoreRowModel,
+    TableOptions,
+  } from "@tanstack/svelte-table";
+
+  type Person = {
+    firstName: string;
+    lastName: string;
+    age: number;
+    visits: number;
+    status: string;
+    progress: number;
+  };
+
+  const defaultData: Person[] = [
+    {
+      firstName: "tanner",
+      lastName: "linsley",
+      age: 24,
+      visits: 100,
+      status: "In Relationship",
+      progress: 50,
+    },
+    {
+      firstName: "tandy",
+      lastName: "miller",
+      age: 40,
+      visits: 40,
+      status: "Single",
+      progress: 80,
+    },
+    {
+      firstName: "joe",
+      lastName: "dirte",
+      age: 45,
+      visits: 20,
+      status: "Complicated",
+      progress: 10,
+    },
+  ];
+
+  const defaultColumns: ColumnDef<Person>[] = [
+    {
+      header: "Name",
+      columns: [
+        {
+          accessorKey: "firstName",
+          cell: (info) => info.getValue(),
+        },
+        {
+          accessorFn: (row) => row.lastName,
+          id: "lastName",
+          cell: (info) => info.getValue(),
+          header: () => "Last Name",
+        },
+      ],
+    },
+    {
+      header: "Info",
+      columns: [
+        {
+          accessorKey: "age",
+          header: () => "Age",
+        },
+        {
+          header: "More Info",
+          columns: [
+            {
+              accessorKey: "visits",
+              header: () => "Visits",
+            },
+            {
+              accessorKey: "status",
+              header: "Status",
+            },
+            {
+              accessorKey: "progress",
+              header: "Profile Progress",
+            },
+          ],
+        },
+      ],
+    },
+  ];
+
+  const options = writable<TableOptions<Person>>({
+    data: defaultData,
+    columns: defaultColumns,
+    getCoreRowModel: getCoreRowModel(),
+  });
+
+  const rerender = () => {
+    options.update((options) => ({
+      ...options,
+      data: defaultData,
+    }));
+  };
+
+  const table = createSvelteTable(options);
+</script>
+
+<div class="p-2">
+  <table>
+    <thead>
+      {#each $table.getHeaderGroups() as headerGroup}
+        <tr>
+          {#each headerGroup.headers as header}
+            <th colSpan={header.colSpan}>
+              {#if !header.isPlaceholder}
+                <svelte:component
+                  this={flexRender(
+                    header.column.columnDef.header,
+                    header.getContext()
+                  )}
+                />
+              {/if}
+            </th>
+          {/each}
+        </tr>
+      {/each}
+    </thead>
+    <tbody>
+      {#each $table.getRowModel().rows as row}
+        <tr>
+          {#each row.getVisibleCells() as cell}
+            <td>
+              <svelte:component
+                this={flexRender(cell.column.columnDef.cell, cell.getContext())}
+              />
+            </td>
+          {/each}
+        </tr>
+      {/each}
+    </tbody>
+  </table>
+  <div class="h-4" />
+  <button on:click={() => rerender()} class="border p-2"> Rerender </button>
+</div>
+
+<style>
+  table {
+    border: 1px solid lightgray;
+  }
+
+  tbody {
+    border-bottom: 1px solid lightgray;
+  }
+
+  th {
+    border-bottom: 1px solid lightgray;
+    border-right: 1px solid lightgray;
+    padding: 2px 4px;
+  }
+
+  tfoot {
+    color: gray;
+  }
+
+  tfoot th {
+    font-weight: normal;
+  }
+</style>
