+++ web-common/src/features/dashboards/stores/metrics-explorer-entity.ts
@@ -136,5 +136,7 @@ export interface MetricsExplorerEntity {
    */
   selectedDimensionName?: string;
 
+  pivot?: boolean;
+
   proto?: string;
 }
