+++ web-common/src/features/dashboards/time-dimension-details/TDDHeader.svelte
@@ -104,7 +104,12 @@
 >
   <div class="flex gap-x-3 items-center font-normal text-gray-500">
     <div class="flex items-center gap-x-2">
-      <div class="flex items-center gap-x-1">
+      <!-- svelte-ignore a11y-click-events-have-key-events -->
+      <!-- Hidden path for opening full pivot -->
+      <div
+        on:click={() => metricsExplorerStore.setPivotMode(metricViewName, true)}
+        class="flex items-center gap-x-1"
+      >
         <Row size="16px" /> Rows
       </div>
