+++ web-common/src/features/dashboards/time-dimension-details/TDDTable.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { createEventDispatcher } from "svelte";
   import { CHECKMARK_COLORS } from "@rilldata/web-common/features/dashboards/config";
-  import Pivot from "@rilldata/web-common/features/dashboards/pivot/Pivot.svelte";
+  import Pivot from "@rilldata/web-common/features/dashboards/pivot/RegularTable.svelte";
   import type {
     PivotPos,
     PivotRenderCallback,
